feat: add stream log example.
