+++ README.md
@@ -129,3 +129,53 @@ See [example/proxy/proxy.go](./example/proxy/proxy.go)
 		},
 	}
 ```
+
+### SSH Stream Log
+
+See [example/stream/stream.go](./example/stream/stream.go)
+
+[embedmd]:# (example/stream/stream.go go /func/ /^}$/)
+```go
+func main() {
+	// Create MakeConfig instance with remote username, server address and path to private key.
+	ssh := &easyssh.MakeConfig{
+		Server:  "localhost",
+		User:    "drone-scp",
+		KeyPath: "./tests/.ssh/id_rsa",
+		Port:    "22",
+		Timeout: 60 * time.Second,
+	}
+
+	// Call Run method with command you want to run on remote server.
+	stdoutChan, stderrChan, doneChan, errChan, err := ssh.Stream("for i in {1..5}; do echo ${i}; sleep 1; done; exit 2;", 60)
+	// Handle errors
+	if err != nil {
+		panic("Can't run remote command: " + err.Error())
+	} else {
+		// read from the output channel until the done signal is passed
+		stillGoing := true
+		isTimeout := true
+		for stillGoing {
+			select {
+			case isTimeout = <-doneChan:
+				stillGoing = false
+			case outline := <-stdoutChan:
+				fmt.Println("out:", outline)
+			case errline := <-stderrChan:
+				fmt.Println("err:", errline)
+			case err = <-errChan:
+			}
+		}
+
+		// get exit code or command error.
+		if err != nil {
+			fmt.Println("err: " + err.Error())
+		}
+
+		// command time out
+		if !isTimeout {
+			fmt.Println("Error: command timeout")
+		}
+	}
+}
+```
